+++ Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/MobileMacPowerError.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public enum MobileMacPowerError: Error, Equatable, Sendable {
+    case unavailable
+    case unsupported
+    case invalidResponse
+}